moved some methods to helpers file, getConstructor takes a tag so that getTag is called less (it is a performance throttle), formatting tweaks
