+++ index.html
@@ -45,7 +45,7 @@
                 <div class="label">Number of iterations:</div>
             </div>
             <div class="input-container">
-                <input class="iterations" type="number" placeholder="1000">
+                <input class="iterations" type="number" placeholder="100">
             </div>
         </div>
         <div class="buttons">
